refact(macros) - Moving macros to AAP branch
